backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
requests
pymysql
python-dotenv
joblib
pandas 
numpy
scikit-learn
email-validator
alembic
pydantic-settings